ttc fixed

+++ sw/airborne/modules/computer_vision/lib/vision/image.h
@@ -65,8 +65,8 @@ struct point_t {
 /* Vector structure for point differences */
 struct flow_t {
   struct point_t pos;         ///< The original position the flow comes from
-  int16_t flow_x;             ///< The x direction flow in subpixels
-  int16_t flow_y;             ///< The y direction flow in subpixels
+  float flow_x;             ///< The x direction flow in subpixels
+  float flow_y;             ///< The y direction flow in subpixels
   uint32_t error;             ///< The matching error in the tracking process
 };
